Fix code placement
