+++ src/components/CategoryPage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const CategoryPage = () => (
+  <div>
+    <h1>Under construction!!</h1>
+  </div>
+);
+
+export default CategoryPage;
