+++ src/redux/Categories/Categories.js
@@ -1,9 +1,1 @@
-import React from 'react';
-
-const Categories = () => (
-  <div>
-    <h1>Under construction!!</h1>
-  </div>
-);
-
-export default Categories;
+// Redux
